# mobile/openapi.yaml
openapi: 3.0.3
info:
  title: ViVu App API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development
  - url: https://api.vivuapp.vn
    description: Production
...
paths:
  /api/v1/auth/register:
    post:
      tags: [Authentication]
      summary: Register (request OTP)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/v1/auth/forget-password:
    post:
      tags: [Authentication]
      summary: Forgot password (request OTP)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgetPasswordRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/v1/auth/verify-otp-register:
    post:
      tags: [Authentication]
      summary: Verify OTP and register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyRequest"
      responses:
        "200":
          description: Auth tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAuthenticationResponse"
  /api/v1/auth/verify-otp-forget-password:
    post:
      tags: [Authentication]
      summary: Verify OTP and reset password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyRequest"
      responses:
        "200":
          description: Auth tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAuthenticationResponse"
  /api/v1/auth/authenticate:
    post:
      tags: [Authentication]
      summary: Authenticate with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthenticationRequest"
      responses:
        "200":
          description: Auth tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAuthenticationResponse"
  /api/v1/auth/refresh-token:
    post:
      tags: [Authentication]
      summary: Refresh access token
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer refresh token
      responses:
        "200":
          description: Raw AuthenticationResponse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationResponse"
...
  /api/v1/posts/{postId}:
    get:
      tags: [Posts]
      security:
        - bearerAuth: []
      summary: Get post by id
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePostResponse"
...
  /api/v1/posts:
    get:
      tags: [Posts]
      security:
        - bearerAuth: []
      summary: Get feed
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePostSlice"
...
  /api/v1/posts/{post_id}:
    delete:
      tags: [Posts]
      security:
        - bearerAuth: []
      summary: Delete post
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
    put:
      tags: [Posts]
      security:
        - bearerAuth: []
      summary: Update post with files
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [postType, visibility]
              properties:
                content:
                  type: string
                postType:
                  type: string
                  enum: [EXPERIENCE, ITINERARY, GROUP_INVITE, TOUR, TOUR_SHARE]
                visibility:
                  type: string
                  enum: [PUBLIC, FRIENDS, PRIVATE]
                tourId:
                  type: string
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePostResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponseBase:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        result:
          nullable: true
    ApiResponseVoid:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
...
    ApiResponseAuthenticationResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            result:
              $ref: "#/components/schemas/AuthenticationResponse"
...
    ApiResponsePostResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            result:
              $ref: "#/components/schemas/PostResponse"
    ApiResponsePostSlice:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            result:
              $ref: "#/components/schemas/SlicePostResponse"
...
    AuthenticationRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RegisterRequest:
      type: object
      required: [fullname, email, username, password]
      properties:
        fullname:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        password:
          type: string
    ForgetPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        newPassword:
          type: string
        confirmPassword:
          type: string
    VerifyRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        otp:
          type: string
    AuthenticationResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
...
    PostResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        content:
          type: string
        postType:
          $ref: "#/components/schemas/PostType"
        visibility:
          $ref: "#/components/schemas/Visibility"
        authorId:
          type: integer
          format: int64
        authorName:
          type: string
        authorAvatarUrl:
          type: string
        mediaUrls:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
        sort:
          $ref: "#/components/schemas/Sort"
        offset:
          type: integer
        paged:
          type: boolean
        unpaged:
          type: boolean
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
...
    SlicePostResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/PostResponse"
        pageable:
          $ref: "#/components/schemas/Pageable"
        size:
          type: integer
        number:
          type: integer
        sort:
          $ref: "#/components/schemas/Sort"
        numberOfElements:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
...
    PostType:
      type: string
      enum: [EXPERIENCE, ITINERARY, GROUP_INVITE, TOUR, TOUR_SHARE]
    Visibility:
      type: string
      enum: [PUBLIC, FRIENDS, PRIVATE]
...